+++ FrontEnd/src/app/core/models/auth-state.ts
@@ -1,7 +1,8 @@
 import { SocialUser } from "angularx-social-login";
+import { Notification } from "./notification";
 
 export interface AuthState {
     isAuthenticated: boolean;
     user: SocialUser | null;
-    errorMessage: string | null;
+    notification: Notification | null;
 }