+++ FrontEnd/src/app/core/models/notification-types.ts
@@ -1,0 +1,7 @@
+export enum NotificationTypes {
+    info = 'info',
+    warning = 'warning',
+    danger = 'danger'
+}
+
+// TODO 'success', 'info', 'warning', 'danger', 'primary', 'secondary', 'light' and 'dark'.