+++ FrontEnd/src/app/core/models/notification.ts
@@ -1,0 +1,9 @@
+import { NotificationTypes } from "./notification-types";
+
+export interface Notification {
+    dismissible: boolean;
+    body: string;
+    type: NotificationTypes
+    // TODO
+    date?: Date;
+}