+++ FrontEnd/src/app/core/store/reducers/auth.reducers.ts
@@ -1,55 +1,64 @@
 import { Action, createReducer, on } from '@ngrx/store';
 
 import { AuthState } from '../../models/auth-state';
+import { NotificationTypes } from '../../models/notification-types';
 import * as AuthActions from '../actions/auth.actions';
 
 export const initialState: AuthState = {
     isAuthenticated: false,
     user: null,
-    errorMessage: null,
+    notification: null
 };
 
 const authReducer = createReducer(
     initialState,
 
-    on(AuthActions.loginStarted, (state) => ({
-        ...state,
+    on(AuthActions.loginStarted, () => ({
         isAuthenticated: false,
         user: null,
-        errorMessage: '',
+        notification: {
+            body: 'Use the Google pop-up to login', //TODO localize
+            type: NotificationTypes.info,
+            dismissible: true
+        }
     })),
 
     on(AuthActions.loginSuccess, (state, { socialUser }) => ({ 
-        ...state,
         isAuthenticated: true,
         user: { ...socialUser },
-        errorMessage: '',
+        notification: null
     })),
     
-    on(AuthActions.loginFail, (state, { errorMessage }) => ({
-        ...state,
+    on(AuthActions.loginFail, () => ({
         isAuthenticated: false,
         user: null,
-        errorMessage,
+        notification: {
+            body: 'Login failed', //TODO localize
+            type: NotificationTypes.danger,
+            dismissible: true
+        }
     })),
     
-    on(AuthActions.loginError, (state, { errorMessage }) => ({
-        ...state,
+    on(AuthActions.loginError, () => ({
         isAuthenticated: false,
         user: null,
-        errorMessage,
+        notification: {
+            body: 'An error occured while logging you in. Please try again later', //TODO localize
+            type: NotificationTypes.danger,
+            dismissible: true
+        }
     })),
     
     on(AuthActions.logoutStarted, () => ({
         isAuthenticated: false,
         user: null,
-        errorMessage: 'by logoutStarted reducer',
+        notification: null
     })),
 
-    on(AuthActions.logoutSuccess, (state) => ({ 
-        ...state,
+    on(AuthActions.logoutSuccess, () => ({ 
+        isAuthenticated: false,
         user: null,
-        errorMessage: ''
+        notification: null
     }))
 );