+++ FrontEnd/src/app/features/managers/components/locations/locations.component.ts
@@ -1,11 +1,12 @@
 import { HttpClient } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 
 @Component({
     selector: 'app-locations',
     templateUrl: './locations.component.html',
     styleUrls: ['./locations.component.scss'],
+    changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LocationsComponent implements OnInit {
     locations$: Observable<string[]>;