+++ FrontEnd/src/app/features/managers/components/managers-main/managers-main.component.ts
@@ -1,9 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-managers-main',
   templateUrl: './managers-main.component.html',
-  styleUrls: ['./managers-main.component.scss']
+  styleUrls: ['./managers-main.component.scss']  ,  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ManagersMainComponent implements OnInit {