+++ FrontEnd/src/app/features/staff/components/staff-main/staff-main.component.ts
@@ -1,9 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-staff-main',
   templateUrl: './staff-main.component.html',
-  styleUrls: ['./staff-main.component.scss']
+  styleUrls: ['./staff-main.component.scss'],    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class StaffMainComponent implements OnInit {