+++ FrontEnd/src/app/shared/shared.module.ts
@@ -1,13 +1,15 @@
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
-import { NavbarComponent } from './components/navbar/navbar.component';
 import { HttpClientModule } from '@angular/common/http';
+import { TranslateModule } from '@ngx-translate/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { SocialLoginModule } from 'angularx-social-login';
+import { AlertModule } from 'ngx-bootstrap/alert';
+
 import { AppRoutingModule } from '../app-routing.module';
-import { TranslateModule } from '@ngx-translate/core';
-import { ToastsComponent } from './components/toasts/toasts.component';
+import { NavbarComponent } from './components/navbar/navbar.component';
+import { NotificationsComponent } from './components/notifications/notifications.component';
 
 /**
  * Contains code (components, directives, and pipes) that will be used across feature modules in the app.
@@ -24,7 +26,7 @@ import { ToastsComponent } from './components/toasts/toasts.component';
 @NgModule({
     declarations: [
         NavbarComponent,
-        ToastsComponent
+        NotificationsComponent
     ],
     imports: [
         BrowserModule,
@@ -34,6 +36,7 @@ import { ToastsComponent } from './components/toasts/toasts.component';
         CommonModule,
         FormsModule,
         ReactiveFormsModule,
+        AlertModule
     ],
     exports: [
         BrowserModule,
@@ -44,7 +47,9 @@ import { ToastsComponent } from './components/toasts/toasts.component';
         CommonModule,
         FormsModule,
         ReactiveFormsModule,
-        NavbarComponent
+        AlertModule,
+        NavbarComponent,
+        NotificationsComponent
     ],
     providers: [],
     bootstrap: [],